requirements.txt
pyudev==0.22.0
six==1.15.0
PyUSB==1.1.0
Zenity==2.0.0